Make the score range from -100 to 100.

+++ portfolio/src/main/java/com/google/sps/data/Message.java
@@ -4,9 +4,9 @@ public final class Message {
   private final long id;
   private final String comment;
   private final long timestamp;
-  private final float score;
+  private final double score;
 
-  public Message(long id, String comment, long timestamp, float score) {
+  public Message(long id, String comment, long timestamp, double score) {
     this.id = id;
     this.comment = comment;
     this.timestamp = timestamp;
